changed from 32 to 8 bit for the Cbc writing

+++ HWInterface/CbcInterface.h
@@ -66,13 +66,13 @@ namespace Ph2_HwInterface
 	        * \param pRegNode : Node of the register to write
 	        * \param pValue : Value to write
 	        */
-	        void WriteCbcReg(Cbc* pCbc, const std::string& pRegNode, uint32_t pValue, bool pVerifLoop = true);
+	        void WriteCbcReg(Cbc* pCbc, const std::string& pRegNode, uint8_t pValue, bool pVerifLoop = true);
 			/*!
 			* \brief Write several registers in both Cbc and Cbc Config File
 			* \param pCbc
 			* \param pVecReq : Vector of pair: Node of the register to write versus value to write
 			*/
-			void WriteCbcMultReg(Cbc* pCbc, std::vector< std::pair<std::string,uint32_t> > pVecReq, bool pVerifLoop = true );
+			void WriteCbcMultReg(Cbc* pCbc, std::vector< std::pair<std::string,uint8_t> > pVecReq, bool pVerifLoop = true );
 			/*!
 			* \brief Write same register in all Cbcs and then UpdateCbc (not working with current Firmware)
 			* \param pModule : Module containing vector of Cbcs
